pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "CHANGE_ME"
authors = [{ name = "Ana", email = "56302061+anafvana@users.noreply.github.com" }]
readme = "README.md"
description = "CHANGE_ME"
dynamic = ["version"]
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
dev = [
    "mypy >= 1.11.1",
    "black >= 24.8.0",
    "pre-commit >= 4.0.1",
    "pyyaml >= 6.0.2",
    "types-PyYAML >= 6.0.2",
]
test = ["pytest >=8.2.2", "pytest-mock >= 3.14.0"]

[tool.flit]
module = "src"

[tool.flit.sdist]
include = ["py.typed"]

[tool.mypy]
python_version = "3.13"
